% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASD_function.R
\name{ASDplot}
\alias{ASDplot}
\title{ASDplot}
\usage{
ASDplot(
  data,
  x,
  y,
  shape,
  methods = shape,
  psize = 5,
  intercept_position = 0.1,
  intercept_color = "black",
  intercept_size = 0.8,
  switch = T,
  title = "",
  subtitle = "",
  ytitle = "Absolute Standardized Difference"
)
}
\arguments{
\item{data}{a dataframe object}

\item{x}{a column includes variables' names to check groups balance (string)}

\item{y}{standaradized differences values (integer)}

\item{shape}{methods used to balance groups differences (factor)}

\item{methods}{a vector of methods names (string)}

\item{psize}{shape size (integer)}

\item{intercept_position}{The cutoff used which indicate significant differences}

\item{intercept_color}{Color of the cutoff bar}

\item{intercept_size}{The size of the cutoff bar}

\item{switch}{Whether to include the cutoff or not (T or F)}

\item{title}{Plot title}

\item{subtitle}{plot subtitle}

\item{ytitle}{Y axis title}
}
\value{
A saved PNG file
}
\description{
ASDplot
}
